requirements.txt
streamlit==1.24.0
python-dotenv==1.0.0
PyMuPDF==1.21.0
requests==2.30.0
beautifulsoup4==4.12.0
google-generativeai==0.1.0  
